Add user ownership of task and update the routes

+++ src/routes/post.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const Post = require('../models/Post');
+const auth = require('../middleware/auth');
 const router = express.Router();
 
 //route for getting all the posts
-router.get('/posts', async (req, res) => {
+router.get('/posts', auth, async (req, res) => {
   try {
     const posts = await Post.find({});
     res.send(posts);
@@ -13,7 +14,7 @@ router.get('/posts', async (req, res) => {
 });
 
 //route for getting a post by id
-router.get('/posts/:id', async (req, res) => {
+router.get('/posts/:id', auth, async (req, res) => {
   try {
     const post = await Post.findById(req.params.id);
     res.send(post);
@@ -23,9 +24,12 @@ router.get('/posts/:id', async (req, res) => {
 });
 
 //route for creating a post
-router.post('/posts', async (req, res) => {
+router.post('/posts', auth, async (req, res) => {
+  const post = new Post({
+    ...req.body,
+    creator: req.user._id,
+  });
   try {
-    const post = new Post(req.body);
     await post.save();
     res.status(201).send(post);
   } catch (error) {
@@ -34,13 +38,44 @@ router.post('/posts', async (req, res) => {
 });
 
 //route for deleting a post
-router.delete('/posts/:id', async (req, res) => {
+router.delete('/posts/:id', auth, async (req, res) => {
+  const userCreator = req.user._id;
   try {
+    const postCreator = await Post.findById(req.params.id);
+    if (!postCreator) {
+      return res.status(404).send();
+    }
+    if (JSON.stringify(postCreator.creator) !== JSON.stringify(userCreator)) {
+      return res.status(500).send('You are not allowed to delete this post');
+    }
     const post = await Post.findByIdAndDelete(req.params.id);
+
     res.send(post);
   } catch (error) {
-    console.log(error);
+    //console.log(error);
     res.status(501).send(error);
+  }
+});
+
+//route for updating a post
+router.patch('/posts/:id', auth, async (req, res) => {
+  const updates = Object.keys(req.body);
+  try {
+    const post = await Post.findById(req.params.id);
+    if (!post) {
+      return res.status(404).send();
+    }
+    if (JSON.stringify(post.creator) != JSON.stringify(req.user._id)) {
+      return res.status(500).send('You are not allowed to update this post');
+    }
+    updates.forEach((update) => {
+      post[update] = req.body[update];
+    });
+    await post.save();
+    res.send(post);
+  } catch (error) {
+    //console.log(error);
+    res.status(500).send(error);
   }
 });
